modularized tts components
